+++ src/types/ReportedMetricsEvents.ts
@@ -1,0 +1,3 @@
+export const BlockRevertedEvent = 'BlockReverted';
+export const FailedTransactionEvent = 'FailedTransaction';
+export const SignatureCollectionErrorEvent = 'SignatureCollectionError';